upgrade to cuda 9

+++ install_cuda.py
@@ -1,29 +1,31 @@
 import os
 
 # Cuda
-os.system('wget https://developer.nvidia.com/compute/cuda/8.0/Prod2/local_installers/cuda-repo-ubuntu1604-8-0-local-ga2_8.0.61-1_amd64-deb -O cuda.deb')
+os.system('wget https://developer.nvidia.com/compute/cuda/9.0/Prod/local_installers/cuda-repo-ubuntu1604-9-0-local_9.0.176-1_amd64-deb -O cuda.deb')
+os.system('sudo apt-key add /var/cuda-repo-<version>/7fa2af80.pub')
 os.system("dpkg -i cuda.deb")
 os.system("apt-get -y update")
 os.system("apt-get -y install cuda")
+os.system("wget https://developer.nvidia.com/compute/cuda/9.0/Prod/patches/1/cuda-repo-ubuntu1604-9-0-local-cublas-performance-update_1.0-1_amd64-deb")
+os.system("dpkg -i cuda-repo-ubuntu1604-9-0-local-cublas-performance-update_1.0-1_amd64-deb")
 
-
-# Cudnn 6
-os.system('wget https://www.dropbox.com/s/w2211kd0u0vgtgs/cudnn-8.0-linux-x64-v6.0.tgz?dl=0 -O cudnn.tgz')
-os.system('rm -rf /opt/cudnn/')
-os.system('mkdir /opt/cudnn/')
-os.system('tar -xvzf cudnn.tgz -C /opt/cudnn/')
+if 0:
+    # Cudnn 6
+    os.system('wget https://www.dropbox.com/s/w2211kd0u0vgtgs/cudnn-8.0-linux-x64-v6.0.tgz?dl=0 -O cudnn.tgz')
+    os.system('rm -rf /opt/cudnn/')
+    os.system('mkdir /opt/cudnn/')
+    os.system('tar -xvzf cudnn.tgz -C /opt/cudnn/')
  
-# Cudnn 5.1
-os.system('wget https://www.dropbox.com/s/12hobmwc1ufxpqr/cudnn-8.0-linux-x64-v5.1.tgz?dl=0 -O cudnn-5.1.tgz')
-os.system('rm -rf /opt/cudnn-5.1/')
-os.system('mkdir /opt/cudnn-5.1/')
-os.system('tar -xvzf cudnn-5.1.tgz -C /opt/cudnn-5.1/')
+    # Cudnn 5.1
+    os.system('wget https://www.dropbox.com/s/12hobmwc1ufxpqr/cudnn-8.0-linux-x64-v5.1.tgz?dl=0 -O cudnn-5.1.tgz')
+    os.system('rm -rf /opt/cudnn-5.1/')
+    os.system('mkdir /opt/cudnn-5.1/')
+    os.system('tar -xvzf cudnn-5.1.tgz -C /opt/cudnn-5.1/')
 
 
-#os.system('mkdir /opt/cudnn/')
-os.system('''echo 'export LD_LIBRARY_PATH=/opt/cudnn/cuda/lib64:$LD_LIBRARY_PATH' >>~/.bashrc''')
-os.system('timedatectl set-timezone Asia/Jerusalem')
-os.system('updatedb &> /dev/null')
+    os.system('''echo 'export LD_LIBRARY_PATH=/opt/cudnn/cuda/lib64:$LD_LIBRARY_PATH' >>~/.bashrc''')
+    os.system('timedatectl set-timezone Asia/Jerusalem')
+    os.system('updatedb &> /dev/null')
 
 
 # Handle nvidia version HELL
